==> scripts/0_download_data.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: '2022-04-27'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(here)
library(FAOSTAT)
library(devtools)
library(doParallel)
library(foreach)

install_github("kdgorospe/fishstatr")

library(fishstatr)
```
# Get list of data

```{r}
fao_codes<-FAOsearch()
```

# Download FAO data

## FishStatJ data

Download from FishStatJ

https://www.fao.org/fishery/statistics-query/en/home

```{r}
fishstat_urls<-c("https://www.fao.org/fishery/static/Data/Capture_2022.1.0.zip", #fisheries production (quantity)
             "https://www.fao.org/fishery/static/Data/Aquaculture_2022.1.0.zip", # aquaculture production (quantity)
             "https://www.fao.org/fishery/static/Data/Aquaculture_2022.1.0.zip", # aquaculture production (value)
       "https://www.fao.org/fishery/static/Data/FI_Trade_2021.1.0.zip", # fish trade aggregated (quantity),
       "https://www.fao.org/fishery/static/Data/FI_Trade_2021.1.0.zip", #fish trade aggregated (value)
            "https://www.fao.org/fishery/static/Data/FI_Trade_Partners_2021.1.0.zip", #fish trade by partner (quantity) 
            "https://www.fao.org/fishery/static/Data/FI_Trade_Partners_2021.1.0.zip", # fish trade by partner (value))
       "https://www.fao.org/fishery/static/Data/GlobalProduction_2022.1.0.zip" #global production by source
)
```

```{r}
cl<-makeCluster(7)
registerDoParallel(cl)
foreach(i = 1:length(fishstat_urls)) %dopar% {
mydir<-here::here("data/raw_data") #path to save files
temp<-tempfile(tmpdir=mydir, fileext=".zip")
download.file(fishstat_urls[i], temp)
unzip(temp, exdir=mydir)
unlink(temp) #delete the zip file
}
```

## Fish Food balance sheets

```{r}
fish_fbs<-get_faostat_bulk(code = "FBS", here("data/raw_data"))
saveRDS(fish_fbs, "data/raw_data/food_balance_all_data.rds")
```

## CN-8 trade codes

## Aus state aquaculture data
