==> scripts/1_tidy_data.Rmd <==
---
title: "tidy_data"
output: html_document
---

#libraries
```{r}

library(tidyverse)
library(janitor)
library(here)
library(countrycode)
library(vroom)
library(raster)
library(sf)
library(readxl)

source(here("src/directories.R"))

```

# Codes

## Country UN Code

```{r}
un_code<-read_csv(here("data/raw_data/CL_FI_COUNTRY_GROUPS.csv")) %>% 
  dplyr::select(UN_Code, ISO3_Code, Name_En)
```
## Species code

```{r}
species_code<-read_csv(here("data/raw_data/CL_FI_SPECIES_GROUPS.csv")) %>% 
  dplyr::select(`3A_Code`, Name_En, Scientific_Name, Major_Group_En, ISSCAAP_Group_En) %>% 
  rename(SPECIES.ALPHA_3_CODE = `3A_Code`,
         species_name = Name_En)
```
## Fishing area code

```{r}
area_code<-read_csv(here("data/raw_data/CL_FI_WATERAREA_GROUPS.csv")) %>% 
  dplyr::select(Code, Name_En, FARegion_Group_En) %>% 
  rename(AREA.CODE = Code,
         Area_name = Name_En)
```
# Data

## Capture fisheries production data

```{r}
#Sort the symbols in values out
## "..." = data unavailable (here taken as zero)
## " " = data not separately available (here taken as zero)
## "-" = nil or zero (make zero)
## "0 0" = more than zero but less than half the unit used (taken to be 0.25)
## F = FAO estimate from available sources of information

# Units are tonnes live weight

capture_prod <- 
   read_csv(here("data", "raw_data", "Capture_Quantity.csv")) %>% 
   rename(UN_Code = COUNTRY.UN_CODE) %>% 
   full_join(., un_code) %>% 
   full_join(., species_code) %>% 
   full_join(., area_code) %>% 
   clean_names() %>% 
   rename(country = name_en,
          species = species_name,
          fishing_area = fa_region_group_en,
          unit = measure,
          year = period) %>% 
  filter(!is.na(value) == T) 

capture_prod<- capture_prod %>% 
  mutate(iso3_code = ifelse(is.na(iso3_code) == T, countrycode(capture_prod$country, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_code)),
         iso3_code = case_when(country == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_code)) %>% 
  filter(!country == "Other nei")

test<-capture_prod %>% 
  filter(is.na(iso3_code) == T)

# test df should have no data

saveRDS(object = capture_prod, file = here("data", "tidy_data", "capture_production_tidy.rds"))
```


## Aquaculture production data

```{r}
aqua_prod_raw <- 
   read_csv(here("data/raw_data/Aquaculture_Quantity.csv")) %>% 
  rename(UN_Code = COUNTRY.UN_CODE) %>% 
   full_join(., un_code) %>% 
   full_join(., species_code) %>% 
   full_join(., area_code) %>% 
  clean_names() %>% 
  rename(country = name_en,
          species = species_name,
          fishing_area = fa_region_group_en,
          unit = measure,
          year = period) %>% 
  filter(!is.na(value) == T)

aqua_prod_raw<- aqua_prod_raw%>% 
  mutate(iso3_code = ifelse(is.na(iso3_code) == T, countrycode(aqua_prod_raw$country, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_code)),
         iso3_code = case_when(country == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_code)) %>% 
  filter(!country == "Other nei")

test<-aqua_prod_raw%>% 
  filter(is.na(iso3_code) == T)

saveRDS(object = aqua_prod_raw, file = here("data", "tidy_data", "aquaculture_production_tidy.rds"))

```

## REVISIT THIS NOT SURE ##

## Fish food balance data
```{r}

fish_fbs_raw <- readRDS(here("data/raw_data/food_balance_all_data.rds")) %>% 
   clean_names()


#remove every other column from row 6

col_seq <- seq(from = 6, to = 118, by=2)

fish_fbs_raw <- fish_fbs_raw[, -col_seq]

(fish_fbs <- fish_fbs_raw %>% 
      pivot_longer(cols = -c(country_name, faostat_group_name, element_name, unit_name), names_to = "year", values_to = "value") %>%
      mutate(year = gsub("x", "", year))
)


saveRDS(object = fish_fbs, file = here("data", "tidy_data", "fish_fbs_tidy.rds"))

```


## Fish trade data - volume, temporal

```{r}
comm_code<- read_csv(here("data/raw_data/CL_FI_COMMODITY_ISSCFC.csv")) %>% 
  dplyr::select(Code, ISSCAAP, Name_En, HS_2017) %>% 
  rename(comm_code = Code,
         Comm_name = Name_En) #commodity code

trade_flow_code<- read_csv(here("data/raw_data/CL_FI_TRADEFLOW_DET.csv")) %>% 
  dplyr::select(Code, Name_En) %>% 
  rename(trade_flow_alpha_code = Code,
         Trade_name = Name_En)#trade flow code
```

```{r}
seafood_trade_raw <- read_csv(here("data", "raw_data", "TRADE_QUANTITY.csv")) %>% 
  rename(comm_code = COMMODITY.FAO_CODE) %>% 
  full_join(., un_code, by = c("COUNTRY_REPORTER.UN_CODE" = "UN_Code")) %>% 
  full_join(., comm_code) %>% 
  full_join(., trade_flow_code, by = c("TRADE_FLOW.ALPHA_CODE" = "trade_flow_alpha_code")) %>% 
   clean_names() %>% 
  rename(country = name_en,
          unit = measure,
          year = period) %>% 
  filter(!is.na(value) == T) 
   

seafood_trade_raw<- seafood_trade_raw%>% 
  mutate(iso3_code = ifelse(is.na(iso3_code) == T, countrycode(seafood_trade_raw$country, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_code)),
         iso3_code = case_when(country == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_code)) %>% 
  filter(!country == "Other nei")

test<-seafood_trade_raw%>% 
  filter(is.na(iso3_code) == T)

saveRDS(object = seafood_trade_raw, file = here("data", "tidy_data", "seafood_trade_volume_product_weight_1976-2019.rds"))
   
```


## Fish trade data - value, temporal
```{r}

seafood_value <- read_csv(here("data", "raw_data", "TRADE_VALUE.csv")) %>% 
   rename(comm_code = COMMODITY.FAO_CODE) %>% 
  full_join(., un_code, by = c("COUNTRY_REPORTER.UN_CODE" = "UN_Code")) %>% 
  full_join(., comm_code) %>% 
  full_join(., trade_flow_code, by = c("TRADE_FLOW.ALPHA_CODE" = "trade_flow_alpha_code")) %>% 
   clean_names() %>% 
    rename(country = name_en,
          unit = measure,
          year = period) %>% 
  filter(!is.na(value) == T) 

seafood_value<- seafood_value%>% 
  mutate(iso3_code = ifelse(is.na(iso3_code) == T, countrycode(seafood_value$country, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_code)),
         iso3_code = case_when(country == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_code)) %>% 
  filter(!country == "Other nei")

test<-seafood_value%>% 
  filter(is.na(iso3_code) == T)

saveRDS(object = seafood_value, file = here("data", "tidy_data", "seafood_trade_value_product_weight_1976-2019.rds"))

```



## Fish trade - volume by trade partner in 2019

```{r}

seafood_trade_volume <- read_csv(here("data", "raw_data", "TRADE_PARTNERS_QUANTITY.csv")) %>% 
  rename(comm_code = COMMODITY.FAO_CODE) %>% 
  full_join(., un_code, by = c("COUNTRY_REPORTER.UN_CODE" = "UN_Code")) %>% 
  rename(iso3_reporter = ISO3_Code,
         country_reporter = Name_En) %>% 
  full_join(., un_code, by = c("COUNTRY_PARTNER.UN_CODE" = "UN_Code")) %>% 
  rename(iso3_partner = ISO3_Code,
         country_partner = Name_En) %>% 
  full_join(., comm_code) %>% 
  full_join(., trade_flow_code, by = c("TRADE_FLOW.ALPHA_CODE" = "trade_flow_alpha_code")) %>% 
   clean_names() %>%  
  rename(unit = measure,
          year = period) %>% 
  filter(!is.na(value) == T) 

seafood_trade_volume<- seafood_trade_volume %>% 
  mutate(iso3_reporter = ifelse(is.na(iso3_reporter) == T, countrycode(seafood_trade_volume$country_reporter, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_reporter)),
         iso3_reporter = case_when(country_reporter == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_reporter),
         iso3_partner = ifelse(is.na(iso3_partner) == T, countrycode(seafood_trade_volume$country_partner, origin = "country.name", destination = "iso3c", warn=TRUE), as.character(iso3_partner)),
         iso3_partner = case_when(country_partner == "Channel Islands" ~ "GB-CHA", TRUE ~ iso3_partner)) %>% 
  filter(!country_reporter == "Other nei",
         !country_partner == "Other nei")

test<-seafood_trade_volume%>% 
  filter(is.na(iso3_reporter) == T,
         is.na(iso3_partner) == T)


saveRDS(object = seafood_trade_volume, file = here("data", "tidy_data", "seafood_trade_volume_product_weight_by_partner_2019.rds"))
```

## CN-8 trade codes

```{r}

raw_codes <- read_csv(here("data", "raw_data", "conversion_factors_cn_8_2001_to_2021.csv")) %>% 
   clean_names() %>% 
   filter(year == 2019) %>% 
   mutate(cn_8 = gsub(" ", "", cn_8))

saveRDS(object = raw_codes, here("data", "tidy_data", "cn_8_codes.rds"))

```


## Full food balance data - tidied to Australia only to speed up read in in other scripts
```{r}

fbs <-readRDS(here("data/raw_data/food_balance_all_data.rds")) %>% 
   clean_names()

fbs_aus <- fbs %>% 
  filter(area == "Australia")

saveRDS(object = fbs_aus, file = here("data", "tidy_data", "food_balance_aus_2010-2019.rds"))
```

## Additional data - currently unused
Seafood trade data from Jessica Gephart for 2018
```{r}

read_csv(here("data", "raw_data", "model0_2018_AUS_JAG.csv")) %>% 
   saveRDS(here("data", "tidy_data", "gephart_australian_imports.rds"))


```
World bank ESG indicators for government effectiveness - taken from World Governance indicators
```{r}

wb_ESG <- read_csv(here("data", "raw_data", "world_bank_ESG_data.csv")) 

gov_eff <- wb_ESG %>% 
   clean_names() %>% 
   filter(series_name == "Government Effectiveness: Estimate") %>% 
   pivot_longer(cols = -c(country_name, country_code, series_name, series_code), names_to = "year", values_to =  "gov_effect") %>% 
   mutate(year = substring(text=year, first = 2, last = 5) %>% 
             as.numeric,
          gov_effect = gov_effect %>% as.numeric) %>% 
   drop_na(gov_effect) %>% 
   select(country_name, country_code, year, gov_effect) %>% 
   rename(iso3c = country_code)

saveRDS(object = gov_eff, file = here("data", "tidy_data", "government_effectiveness.rds"))

```

Global slavery index
```{r}

read_csv(file = here("data", "raw_data", "gsi.csv")) %>% 
   saveRDS(file = here("data", "tidy_data", "gsi.rds"))

```

Environmental Performance Index
```{r}

epi_2018 <- read_csv(here("data", "raw_data", "epi_2018.csv"))
epi_2020 <- read_csv(here("data", "raw_data", "epi_2020.csv"))

ecosystem_viability <- epi_2020 %>% select(iso, country, ECO.new) %>% 
   rename(iso3c = iso)

saveRDS(object = ecosystem_viability, file = here("data", "tidy_data", "ecosystem_vaibility.rds"))


```


Spatialised Australian landings raster
```{r}
wgs_84 <- "EPSG:4236"

aus_spatial <- readRDS(here("data/raw_data/aus_landings.rds")) %>% 
   group_by(Cell, LatCentre, LonCentre) %>% 
   summarise(reported = sum(Reported)) %>% 
   st_as_sf(coords = c("LonCentre", "LatCentre"))

raster_0.5 <- raster(crs=wgs_84, res=0.5)
points <- st_coordinates(aus_spatial$geometry)


aus_fishing <- rasterize(points, raster_0.5, field = aus_spatial$reported, fun = sum)

plot(aus_fishing)
writeRaster(x=aus_fishing, filename = here("data/spatial/aus_industrial_fishing.tif"), overwrite=TRUE)

```


Tidy Aus state level aquaculture data
```{r}
aus_state_aqua <- read_csv(here("data/raw_data/aus_aquaculture_production.csv")) %>%
  pivot_longer(names_to = "year", values_to = "quantity", -c(state, species)) %>% 
  mutate(quantity = gsub(",", "", quantity)) %>% 
  mutate(year = as.integer(year),
         quantity = as.double(quantity))


saveRDS(aus_state_aqua, here("data/tidy_data/state_aquaculture_production_tidy.rds"))


```
