==> scripts/2_convert_product_to_live_equivalents.Rmd <==
---
title: "Live equivalents conversion"
output: html_document
---

libraries

```{r}

library(tidyverse)
library(janitor)
library(here)
library(stringr)
```


# Data


## Trade by partner

```{r}
seafood_trade_products <- readRDS(here("data", "tidy_data", "seafood_trade_volume_product_weight_by_partner_2019.rds")) %>% 
  filter(country_reporter == "Australia") %>% 
  mutate(hs_code = stringr::str_replace_all(string = hs_2017, pattern = "[[:punct:]]", replacement = ""))


hs_codes <- readRDS(file = here("data", "tidy_data", "cn_8_codes.rds")) %>% 
  mutate(cn_8 = substring(text = cn_8, first = 1, last=6)) %>% 
  group_by(cn_8) %>% 
  summarise(mean_cf = mean(cf),
            median_cf = median(cf))


seafood_live_equivalents <- 
  seafood_trade_products %>% 
  left_join(y=hs_codes %>%  rename(hs_code = cn_8), by = c("hs_code")) %>% 
  mutate(value_LE_med_cf = if_else(median_cf==0,true = 0, false = value*median_cf),
         value_LE_mean_cf = if_else(mean_cf==0,true = 0, false = value*mean_cf)) %>% 
  distinct()

saveRDS(object = seafood_live_equivalents, file = here("data", "tidy_data", "seafood_trade_volume_live_equivalent_by_partner_2019.rds"))


seafood_live_equivalents %>% filter(country_partner=="Thailand")

seafood_trade_products %>% filter(country_partner=="Thailand" & trade_name == "Imports")


```
